Move build.py and send.py into thumby.py which now accepts a parameter
